factor out input binder
